+++ flutter_navigation_generator_annotations/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added methodName and updated routeName to include web parameters
 - Added unknownRoute to handle unknown routes
 - Added guards to prevent navigation under certain conditions
+- Added defaultGuards to apply to all routes where no guards are specified
 
 ## 1.0.1 (2023-11-14)